feature(bento-pan-zoom): revert gesture refactor
